pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "addict>=2.4.0",
    "aiomysql>=0.3.2",
    "aiosqlite>=0.21.0",
    "cryptography>=46.0.3",
    "fake-useragent>=2.2.0",
    "fastapi>=0.119.1",
    "fastapi-cdn-host>=0.9.2",
    "loguru>=0.7.3",
    "modelscope[multi-modal]>=1.31.0",
    "notebook>=7.4.7",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pyjwt>=2.10.1",
    "snowflake-id>=1.0.2",
    "sqlmodel>=0.0.27",
    "ultralytics>=8.3.228",
    "uvicorn[standard]>=0.38.0",
    "tf-keras>=2.20.1",
    "async-class>=0.5.0",
    "pybrowsers>=1.3.1",
    "chrome-fingerprints>=1.1",
    "hcaptcha-challenger>=0.18.13",
    "recognizer>=1.4",
    "browserforge[all]>=1.2.3",
    "ipywidgets>=8.1.8",
    "ipykernel>=7.1.0",
    "patchright>=1.56.0",
    "greenlet>=3.3.0",
    "jupyterlab>=4.5.0",
    "onnxruntime>=1.23.2",
    "aiofiles>=25.1.0",
    "websockets>=15.0.1",
    "aiortc>=1.14.0",
    "apscheduler>=3.10.4",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
